Camera Link fixes for Linux GCC 4
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@465 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_epix_xclib.c
@@ -1013,7 +1013,7 @@ mxd_epix_xclib_get_frame( MX_VIDEO_INPUT *vinput, MX_IMAGE_FRAME **frame )
 	MX_DEBUG(-2,("%s: (*frame)->image_data = %p",
 		fname, (*frame)->image_data));
 	MX_DEBUG(-2,("%s: (*frame)->image_length = %lu, bytes_per_image = %lu",
-		fname, (*frame)->image_length, bytes_per_image));
+	    fname, (unsigned long) (*frame)->image_length, bytes_per_image));
 #endif
 
 	if ( ( (*frame)->image_data != NULL )
